+++ packages/orq-rc/RELEASES.md
@@ -2929,3 +2929,13 @@ Based on:
 - [typescript v3.11.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v3.11.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.3 - packages/orq-rc
+
+## 2025-08-11 09:43:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.600.2 (2.677.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.11.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v3.11.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.4 - packages/orq-rc